refactor(carver): adjust file structure and delete useless files
