+++ carver/src/main/java/com/kokomi/carver/view/setting/Color.kt
@@ -1,4 +1,4 @@
-package com.kokomi.carver.ui.setting
+package com.kokomi.carver.view.setting
 
 import androidx.compose.ui.graphics.Color
