+++ carver/src/main/java/com/kokomi/carver/view/setting/Type.kt
@@ -1,4 +1,4 @@
-package com.kokomi.carver.ui.setting
+package com.kokomi.carver.view.setting
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
