+++ carver/src/main/res/layout/activity_carver.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/app_black"
-    tools:context=".ui.capture.CarverActivity">
+    tools:context=".view.carver.CarverActivity">
 
     <TextView
         android:id="@+id/tv_carver_status_bar"
